Create UI folder in SynthesisCore and create ToolbarTools
